requirements.txt
attrs==19.3.0
click==7.1.2
coverage==5.2.1
Flask==1.1.2
Flask-Caching==1.9.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.0
python-dotenv==0.14.0
redis==3.5.3
six==1.15.0
wcwidth==0.2.5
webargs==6.1.0
Werkzeug==1.0.1
